pyproject.toml
[project]
name = "mcp-reddit-mod"
version = "0.1.0"
description = "AI-assisted Reddit moderation MCP server"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.26.0",
    "praw>=7.8.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
mcp-reddit-mod = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-anyio>=0.0.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.8.0",
    "mypy>=1.8.0",
    "vulture>=2.15",
    "pre-commit>=4.5",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "praw.*",
    "dotenv.*",
]
ignore_missing_imports = true

[tool.vulture]
min_confidence = 80
paths = ["src"]
exclude = ["tests/", ".venv/", "node_modules/"]
ignore_names = [
    # Common false positives
    "visit_*",
    "asdict",
    "do_.*",
    # Test fixtures
    "test_*",
    # FastMCP decorators may create false positives
    "_.*",
]
